Module Import.Ast_builder

include module type of struct include Ppxlib.Ast_builder.Default end

module Located = Ppxlib.Ast_builder.Default.Located

val attribute : loc:Astlib.Location.t -> name:string Astlib.Location.loc -> payload:Astlib.Ast_500.Parsetree.payload -> Astlib.Ast_500.Parsetree.attribute

val binding_op : loc:Astlib.Location.t -> op:string Astlib.Location.loc -> pat:Astlib.Ast_500.Parsetree.pattern -> exp:Astlib.Ast_500.Parsetree.expression -> Astlib.Ast_500.Parsetree.binding_op

val case : lhs:Astlib.Ast_500.Parsetree.pattern -> guard:Astlib.Ast_500.Parsetree.expression option -> rhs:Astlib.Ast_500.Parsetree.expression -> Astlib.Ast_500.Parsetree.case

val pcl_constr : loc:Astlib.Location.t -> Astlib.Longident.t Astlib.Location.loc -> Astlib.Ast_500.Parsetree.core_type list -> Astlib.Ast_500.Parsetree.class_expr

val pcl_structure : loc:Astlib.Location.t -> Astlib.Ast_500.Parsetree.class_structure -> Astlib.Ast_500.Parsetree.class_expr

val pcl_fun : loc:Astlib.Location.t -> Astlib.Ast_500.Asttypes.arg_label -> Astlib.Ast_500.Parsetree.expression option -> Astlib.Ast_500.Parsetree.pattern -> Astlib.Ast_500.Parsetree.class_expr -> Astlib.Ast_500.Parsetree.class_expr

val pcl_apply : loc:Astlib.Location.t -> Astlib.Ast_500.Parsetree.class_expr -> (Astlib.Ast_500.Asttypes.arg_label * Astlib.Ast_500.Parsetree.expression) list -> Astlib.Ast_500.Parsetree.class_expr

val pcl_let : loc:Astlib.Location.t -> Astlib.Ast_500.Asttypes.rec_flag -> Astlib.Ast_500.Parsetree.value_binding list -> Astlib.Ast_500.Parsetree.class_expr -> Astlib.Ast_500.Parsetree.class_expr

val pcl_constraint : loc:Astlib.Location.t -> Astlib.Ast_500.Parsetree.class_expr -> Astlib.Ast_500.Parsetree.class_type -> Astlib.Ast_500.Parsetree.class_expr

val pcl_extension : loc:Astlib.Location.t -> (string Astlib.Location.loc * Astlib.Ast_500.Parsetree.payload) -> Astlib.Ast_500.Parsetree.class_expr

val pcl_open : loc:Astlib.Location.t -> Astlib.Longident.t Astlib.Location.loc Astlib.Ast_500.Parsetree.open_infos -> Astlib.Ast_500.Parsetree.class_expr -> Astlib.Ast_500.Parsetree.class_expr

val pcf_inherit : loc:Astlib.Location.t -> Astlib.Ast_500.Asttypes.override_flag -> Astlib.Ast_500.Parsetree.class_expr -> string Astlib.Location.loc option -> Astlib.Ast_500.Parsetree.class_field

val pcf_val : loc:Astlib.Location.t -> (string Astlib.Location.loc * Astlib.Ast_500.Asttypes.mutable_flag * Astlib.Ast_500.Parsetree.class_field_kind) -> Astlib.Ast_500.Parsetree.class_field

val pcf_method : loc:Astlib.Location.t -> (string Astlib.Location.loc * Astlib.Ast_500.Asttypes.private_flag * Astlib.Ast_500.Parsetree.class_field_kind) -> Astlib.Ast_500.Parsetree.class_field

val pcf_constraint : loc:Astlib.Location.t -> (Astlib.Ast_500.Parsetree.core_type * Astlib.Ast_500.Parsetree.core_type) -> Astlib.Ast_500.Parsetree.class_field

val pcf_initializer : loc:Astlib.Location.t -> Astlib.Ast_500.Parsetree.expression -> Astlib.Ast_500.Parsetree.class_field

val pcf_attribute : loc:Astlib.Location.t -> Astlib.Ast_500.Parsetree.attribute -> Astlib.Ast_500.Parsetree.class_field

val pcf_extension : loc:Astlib.Location.t -> (string Astlib.Location.loc * Astlib.Ast_500.Parsetree.payload) -> Astlib.Ast_500.Parsetree.class_field

val class_infos : loc:Astlib.Location.t -> virt:Astlib.Ast_500.Asttypes.virtual_flag -> params: (Astlib.Ast_500.Parsetree.core_type * (Astlib.Ast_500.Asttypes.variance * Astlib.Ast_500.Asttypes.injectivity)) list -> name:string Astlib.Location.loc -> expr:'a -> 'a Astlib.Ast_500.Parsetree.class_infos

val class_signature : self:Astlib.Ast_500.Parsetree.core_type -> fields:Astlib.Ast_500.Parsetree.class_type_field list -> Astlib.Ast_500.Parsetree.class_signature

val class_structure : self:Astlib.Ast_500.Parsetree.pattern -> fields:Astlib.Ast_500.Parsetree.class_field list -> Astlib.Ast_500.Parsetree.class_structure

val pcty_constr : loc:Astlib.Location.t -> Astlib.Longident.t Astlib.Location.loc -> Astlib.Ast_500.Parsetree.core_type list -> Astlib.Ast_500.Parsetree.class_type

val pcty_signature : loc:Astlib.Location.t -> Astlib.Ast_500.Parsetree.class_signature -> Astlib.Ast_500.Parsetree.class_type

val pcty_arrow : loc:Astlib.Location.t -> Astlib.Ast_500.Asttypes.arg_label -> Astlib.Ast_500.Parsetree.core_type -> Astlib.Ast_500.Parsetree.class_type -> Astlib.Ast_500.Parsetree.class_type

val pcty_extension : loc:Astlib.Location.t -> (string Astlib.Location.loc * Astlib.Ast_500.Parsetree.payload) -> Astlib.Ast_500.Parsetree.class_type

val pcty_open : loc:Astlib.Location.t -> Astlib.Longident.t Astlib.Location.loc Astlib.Ast_500.Parsetree.open_infos -> Astlib.Ast_500.Parsetree.class_type -> Astlib.Ast_500.Parsetree.class_type

val pctf_inherit : loc:Astlib.Location.t -> Astlib.Ast_500.Parsetree.class_type -> Astlib.Ast_500.Parsetree.class_type_field

val pctf_val : loc:Astlib.Location.t -> (string Astlib.Location.loc * Astlib.Ast_500.Asttypes.mutable_flag * Astlib.Ast_500.Asttypes.virtual_flag * Astlib.Ast_500.Parsetree.core_type) -> Astlib.Ast_500.Parsetree.class_type_field

val pctf_method : loc:Astlib.Location.t -> (string Astlib.Location.loc * Astlib.Ast_500.Asttypes.private_flag * Astlib.Ast_500.Asttypes.virtual_flag * Astlib.Ast_500.Parsetree.core_type) -> Astlib.Ast_500.Parsetree.class_type_field

val pctf_constraint : loc:Astlib.Location.t -> (Astlib.Ast_500.Parsetree.core_type * Astlib.Ast_500.Parsetree.core_type) -> Astlib.Ast_500.Parsetree.class_type_field

val pctf_attribute : loc:Astlib.Location.t -> Astlib.Ast_500.Parsetree.attribute -> Astlib.Ast_500.Parsetree.class_type_field

val pctf_extension : loc:Astlib.Location.t -> (string Astlib.Location.loc * Astlib.Ast_500.Parsetree.payload) -> Astlib.Ast_500.Parsetree.class_type_field

val comprehension : body:Astlib.Ast_500.Parsetree.expression -> clauses:Astlib.Ast_500.Parsetree.comprehension_clause list -> Astlib.Ast_500.Parsetree.comprehension

val comprehension_clause_binding : cb_pattern:Astlib.Ast_500.Parsetree.pattern -> cb_iterator:Astlib.Ast_500.Parsetree.comprehension_iterator -> cb_attributes:Astlib.Ast_500.Parsetree.attribute list -> Astlib.Ast_500.Parsetree.comprehension_clause_binding

val constructor_argument : loc:Astlib.Location.t -> modalities:Astlib.Ast_500.Parsetree.modality Astlib.Location.loc list -> type_:Astlib.Ast_500.Parsetree.core_type -> Astlib.Ast_500.Parsetree.constructor_argument

val ptyp_constr : loc:Astlib.Location.t -> Astlib.Longident.t Astlib.Location.loc -> Astlib.Ast_500.Parsetree.core_type list -> Astlib.Ast_500.Parsetree.core_type

val ptyp_object : loc:Astlib.Location.t -> Astlib.Ast_500.Parsetree.object_field list -> Astlib.Ast_500.Asttypes.closed_flag -> Astlib.Ast_500.Parsetree.core_type

val ptyp_class : loc:Astlib.Location.t -> Astlib.Longident.t Astlib.Location.loc -> Astlib.Ast_500.Parsetree.core_type list -> Astlib.Ast_500.Parsetree.core_type

val ptyp_variant : loc:Astlib.Location.t -> Astlib.Ast_500.Parsetree.row_field list -> Astlib.Ast_500.Asttypes.closed_flag -> string list option -> Astlib.Ast_500.Parsetree.core_type

val ptyp_package : loc:Astlib.Location.t -> (Astlib.Longident.t Astlib.Location.loc * (Astlib.Longident.t Astlib.Location.loc * Astlib.Ast_500.Parsetree.core_type) list) -> Astlib.Ast_500.Parsetree.core_type

val ptyp_of_kind : loc:Astlib.Location.t -> Astlib.Ast_500.Parsetree.jkind_annotation -> Astlib.Ast_500.Parsetree.core_type

val ptyp_extension : loc:Astlib.Location.t -> (string Astlib.Location.loc * Astlib.Ast_500.Parsetree.payload) -> Astlib.Ast_500.Parsetree.core_type

val pdir_string : loc:Astlib.Location.t -> string -> Astlib.Ast_500.Parsetree.directive_argument

val pdir_int : loc:Astlib.Location.t -> string -> char option -> Astlib.Ast_500.Parsetree.directive_argument

val pdir_ident : loc:Astlib.Location.t -> Astlib.Longident.t -> Astlib.Ast_500.Parsetree.directive_argument

val pdir_bool : loc:Astlib.Location.t -> bool -> Astlib.Ast_500.Parsetree.directive_argument

val pexp_ident : loc:Astlib.Location.t -> Astlib.Longident.t Astlib.Location.loc -> Astlib.Ast_500.Parsetree.expression

val pexp_constant : loc:Astlib.Location.t -> Astlib.Ast_500.Parsetree.constant -> Astlib.Ast_500.Parsetree.expression

val pexp_match : loc:Astlib.Location.t -> Astlib.Ast_500.Parsetree.expression -> Astlib.Ast_500.Parsetree.case list -> Astlib.Ast_500.Parsetree.expression

val pexp_try : loc:Astlib.Location.t -> Astlib.Ast_500.Parsetree.expression -> Astlib.Ast_500.Parsetree.case list -> Astlib.Ast_500.Parsetree.expression

val pexp_construct : loc:Astlib.Location.t -> Astlib.Longident.t Astlib.Location.loc -> Astlib.Ast_500.Parsetree.expression option -> Astlib.Ast_500.Parsetree.expression

val pexp_variant : loc:Astlib.Location.t -> string -> Astlib.Ast_500.Parsetree.expression option -> Astlib.Ast_500.Parsetree.expression

val pexp_record : loc:Astlib.Location.t -> (Astlib.Longident.t Astlib.Location.loc * Astlib.Ast_500.Parsetree.expression) list -> Astlib.Ast_500.Parsetree.expression option -> Astlib.Ast_500.Parsetree.expression

val pexp_field : loc:Astlib.Location.t -> Astlib.Ast_500.Parsetree.expression -> Astlib.Longident.t Astlib.Location.loc -> Astlib.Ast_500.Parsetree.expression

val pexp_setfield : loc:Astlib.Location.t -> Astlib.Ast_500.Parsetree.expression -> Astlib.Longident.t Astlib.Location.loc -> Astlib.Ast_500.Parsetree.expression -> Astlib.Ast_500.Parsetree.expression

val pexp_ifthenelse : loc:Astlib.Location.t -> Astlib.Ast_500.Parsetree.expression -> Astlib.Ast_500.Parsetree.expression -> Astlib.Ast_500.Parsetree.expression option -> Astlib.Ast_500.Parsetree.expression

val pexp_sequence : loc:Astlib.Location.t -> Astlib.Ast_500.Parsetree.expression -> Astlib.Ast_500.Parsetree.expression -> Astlib.Ast_500.Parsetree.expression

val pexp_while : loc:Astlib.Location.t -> Astlib.Ast_500.Parsetree.expression -> Astlib.Ast_500.Parsetree.expression -> Astlib.Ast_500.Parsetree.expression

val pexp_for : loc:Astlib.Location.t -> Astlib.Ast_500.Parsetree.pattern -> Astlib.Ast_500.Parsetree.expression -> Astlib.Ast_500.Parsetree.expression -> Astlib.Ast_500.Asttypes.direction_flag -> Astlib.Ast_500.Parsetree.expression -> Astlib.Ast_500.Parsetree.expression

val pexp_coerce : loc:Astlib.Location.t -> Astlib.Ast_500.Parsetree.expression -> Astlib.Ast_500.Parsetree.core_type option -> Astlib.Ast_500.Parsetree.core_type -> Astlib.Ast_500.Parsetree.expression

val pexp_send : loc:Astlib.Location.t -> Astlib.Ast_500.Parsetree.expression -> string Astlib.Location.loc -> Astlib.Ast_500.Parsetree.expression

val pexp_new : loc:Astlib.Location.t -> Astlib.Longident.t Astlib.Location.loc -> Astlib.Ast_500.Parsetree.expression

val pexp_setvar : loc:Astlib.Location.t -> string Astlib.Location.loc -> Astlib.Ast_500.Parsetree.expression -> Astlib.Ast_500.Parsetree.expression

val pexp_override : loc:Astlib.Location.t -> (string Astlib.Location.loc * Astlib.Ast_500.Parsetree.expression) list -> Astlib.Ast_500.Parsetree.expression

val pexp_letmodule : loc:Astlib.Location.t -> string option Astlib.Location.loc -> Astlib.Ast_500.Parsetree.module_expr -> Astlib.Ast_500.Parsetree.expression -> Astlib.Ast_500.Parsetree.expression

val pexp_letexception : loc:Astlib.Location.t -> Astlib.Ast_500.Parsetree.extension_constructor -> Astlib.Ast_500.Parsetree.expression -> Astlib.Ast_500.Parsetree.expression

val pexp_assert : loc:Astlib.Location.t -> Astlib.Ast_500.Parsetree.expression -> Astlib.Ast_500.Parsetree.expression

val pexp_lazy : loc:Astlib.Location.t -> Astlib.Ast_500.Parsetree.expression -> Astlib.Ast_500.Parsetree.expression

val pexp_poly : loc:Astlib.Location.t -> Astlib.Ast_500.Parsetree.expression -> Astlib.Ast_500.Parsetree.core_type option -> Astlib.Ast_500.Parsetree.expression

val pexp_object : loc:Astlib.Location.t -> Astlib.Ast_500.Parsetree.class_structure -> Astlib.Ast_500.Parsetree.expression

val pexp_pack : loc:Astlib.Location.t -> Astlib.Ast_500.Parsetree.module_expr -> Astlib.Ast_500.Parsetree.expression

val pexp_open : loc:Astlib.Location.t -> Astlib.Ast_500.Parsetree.module_expr Astlib.Ast_500.Parsetree.open_infos -> Astlib.Ast_500.Parsetree.expression -> Astlib.Ast_500.Parsetree.expression

val pexp_letop : loc:Astlib.Location.t -> Astlib.Ast_500.Parsetree.letop -> Astlib.Ast_500.Parsetree.expression

val pexp_extension : loc:Astlib.Location.t -> (string Astlib.Location.loc * Astlib.Ast_500.Parsetree.payload) -> Astlib.Ast_500.Parsetree.expression

val pexp_unreachable : loc:Astlib.Location.t -> Astlib.Ast_500.Parsetree.expression

val pexp_stack : loc:Astlib.Location.t -> Astlib.Ast_500.Parsetree.expression -> Astlib.Ast_500.Parsetree.expression

val pexp_comprehension : loc:Astlib.Location.t -> Astlib.Ast_500.Parsetree.comprehension_expression -> Astlib.Ast_500.Parsetree.expression

val pexp_overwrite : loc:Astlib.Location.t -> Astlib.Ast_500.Parsetree.expression -> Astlib.Ast_500.Parsetree.expression -> Astlib.Ast_500.Parsetree.expression

val pexp_quote : loc:Astlib.Location.t -> Astlib.Ast_500.Parsetree.expression -> Astlib.Ast_500.Parsetree.expression

val pexp_splice : loc:Astlib.Location.t -> Astlib.Ast_500.Parsetree.expression -> Astlib.Ast_500.Parsetree.expression

val pexp_hole : loc:Astlib.Location.t -> Astlib.Ast_500.Parsetree.expression

val extension_constructor : loc:Astlib.Location.t -> name:string Astlib.Location.loc -> kind:Astlib.Ast_500.Parsetree.extension_constructor_kind -> Astlib.Ast_500.Parsetree.extension_constructor

val function_constraint : mode_annotations:Astlib.Ast_500.Parsetree.mode Astlib.Location.loc list -> ret_mode_annotations:Astlib.Ast_500.Parsetree.mode Astlib.Location.loc list -> ret_type_constraint:Astlib.Ast_500.Parsetree.type_constraint option -> Astlib.Ast_500.Parsetree.function_constraint

val pparam_val : loc:Astlib.Location.t -> Astlib.Ast_500.Asttypes.arg_label -> Astlib.Ast_500.Parsetree.expression option -> Astlib.Ast_500.Parsetree.pattern -> Astlib.Ast_500.Parsetree.function_param

val pparam_newtype : loc:Astlib.Location.t -> string Astlib.Location.loc -> Astlib.Ast_500.Parsetree.jkind_annotation option -> Astlib.Ast_500.Parsetree.function_param

val pjk_default : loc:Astlib.Location.t -> Astlib.Ast_500.Parsetree.jkind_annotation

val pjk_abbreviation : loc:Astlib.Location.t -> string -> Astlib.Ast_500.Parsetree.jkind_annotation

val pjk_mod : loc:Astlib.Location.t -> Astlib.Ast_500.Parsetree.jkind_annotation -> Astlib.Ast_500.Parsetree.mode Astlib.Location.loc list -> Astlib.Ast_500.Parsetree.jkind_annotation

val pjk_with : loc:Astlib.Location.t -> Astlib.Ast_500.Parsetree.jkind_annotation -> Astlib.Ast_500.Parsetree.core_type -> Astlib.Ast_500.Parsetree.modality Astlib.Location.loc list -> Astlib.Ast_500.Parsetree.jkind_annotation

val pjk_kind_of : loc:Astlib.Location.t -> Astlib.Ast_500.Parsetree.core_type -> Astlib.Ast_500.Parsetree.jkind_annotation

val pjk_product : loc:Astlib.Location.t -> Astlib.Ast_500.Parsetree.jkind_annotation list -> Astlib.Ast_500.Parsetree.jkind_annotation

val letop : let_:Astlib.Ast_500.Parsetree.binding_op -> ands:Astlib.Ast_500.Parsetree.binding_op list -> body:Astlib.Ast_500.Parsetree.expression -> Astlib.Ast_500.Parsetree.letop

val location : start:Lexing.position -> end_:Lexing.position -> ghost:bool -> Astlib.Location.t

val module_binding : loc:Astlib.Location.t -> name:string option Astlib.Location.loc -> expr:Astlib.Ast_500.Parsetree.module_expr -> Astlib.Ast_500.Parsetree.module_binding

val pmod_ident : loc:Astlib.Location.t -> Astlib.Longident.t Astlib.Location.loc -> Astlib.Ast_500.Parsetree.module_expr

val pmod_structure : loc:Astlib.Location.t -> Astlib.Ast_500.Parsetree.structure_item list -> Astlib.Ast_500.Parsetree.module_expr

val pmod_functor : loc:Astlib.Location.t -> Astlib.Ast_500.Parsetree.functor_parameter -> Astlib.Ast_500.Parsetree.module_expr -> Astlib.Ast_500.Parsetree.module_expr

val pmod_apply : loc:Astlib.Location.t -> Astlib.Ast_500.Parsetree.module_expr -> Astlib.Ast_500.Parsetree.module_expr -> Astlib.Ast_500.Parsetree.module_expr

val pmod_unpack : loc:Astlib.Location.t -> Astlib.Ast_500.Parsetree.expression -> Astlib.Ast_500.Parsetree.module_expr

val pmod_extension : loc:Astlib.Location.t -> (string Astlib.Location.loc * Astlib.Ast_500.Parsetree.payload) -> Astlib.Ast_500.Parsetree.module_expr

val pmod_instance : loc:Astlib.Location.t -> Astlib.Ast_500.Parsetree.module_instance -> Astlib.Ast_500.Parsetree.module_expr

val module_instance : instance_head:string -> instance_args:(string * Astlib.Ast_500.Parsetree.module_instance) list -> Astlib.Ast_500.Parsetree.module_instance

val module_substitution : loc:Astlib.Location.t -> name:string Astlib.Location.loc -> manifest:Astlib.Longident.t Astlib.Location.loc -> Astlib.Ast_500.Parsetree.module_substitution

val pmty_ident : loc:Astlib.Location.t -> Astlib.Longident.t Astlib.Location.loc -> Astlib.Ast_500.Parsetree.module_type

val pmty_with : loc:Astlib.Location.t -> Astlib.Ast_500.Parsetree.module_type -> Astlib.Ast_500.Parsetree.with_constraint list -> Astlib.Ast_500.Parsetree.module_type

val pmty_typeof : loc:Astlib.Location.t -> Astlib.Ast_500.Parsetree.module_expr -> Astlib.Ast_500.Parsetree.module_type

val pmty_extension : loc:Astlib.Location.t -> (string Astlib.Location.loc * Astlib.Ast_500.Parsetree.payload) -> Astlib.Ast_500.Parsetree.module_type

val pmty_alias : loc:Astlib.Location.t -> Astlib.Longident.t Astlib.Location.loc -> Astlib.Ast_500.Parsetree.module_type

val pmty_strengthen : loc:Astlib.Location.t -> Astlib.Ast_500.Parsetree.module_type -> Astlib.Longident.t Astlib.Location.loc -> Astlib.Ast_500.Parsetree.module_type

val module_type_declaration : loc:Astlib.Location.t -> name:string Astlib.Location.loc -> type_:Astlib.Ast_500.Parsetree.module_type option -> Astlib.Ast_500.Parsetree.module_type_declaration

val otag : loc:Astlib.Location.t -> string Astlib.Location.loc -> Astlib.Ast_500.Parsetree.core_type -> Astlib.Ast_500.Parsetree.object_field

val oinherit : loc:Astlib.Location.t -> Astlib.Ast_500.Parsetree.core_type -> Astlib.Ast_500.Parsetree.object_field

val open_infos : loc:Astlib.Location.t -> expr:'a -> override:Astlib.Ast_500.Asttypes.override_flag -> 'a Astlib.Ast_500.Parsetree.open_infos

val ppat_any : loc:Astlib.Location.t -> Astlib.Ast_500.Parsetree.pattern

val ppat_var : loc:Astlib.Location.t -> string Astlib.Location.loc -> Astlib.Ast_500.Parsetree.pattern

val ppat_alias : loc:Astlib.Location.t -> Astlib.Ast_500.Parsetree.pattern -> string Astlib.Location.loc -> Astlib.Ast_500.Parsetree.pattern

val ppat_constant : loc:Astlib.Location.t -> Astlib.Ast_500.Parsetree.constant -> Astlib.Ast_500.Parsetree.pattern

val ppat_interval : loc:Astlib.Location.t -> Astlib.Ast_500.Parsetree.constant -> Astlib.Ast_500.Parsetree.constant -> Astlib.Ast_500.Parsetree.pattern

val ppat_variant : loc:Astlib.Location.t -> string -> Astlib.Ast_500.Parsetree.pattern option -> Astlib.Ast_500.Parsetree.pattern

val ppat_record : loc:Astlib.Location.t -> (Astlib.Longident.t Astlib.Location.loc * Astlib.Ast_500.Parsetree.pattern) list -> Astlib.Ast_500.Asttypes.closed_flag -> Astlib.Ast_500.Parsetree.pattern

val ppat_or : loc:Astlib.Location.t -> Astlib.Ast_500.Parsetree.pattern -> Astlib.Ast_500.Parsetree.pattern -> Astlib.Ast_500.Parsetree.pattern

val ppat_type : loc:Astlib.Location.t -> Astlib.Longident.t Astlib.Location.loc -> Astlib.Ast_500.Parsetree.pattern

val ppat_lazy : loc:Astlib.Location.t -> Astlib.Ast_500.Parsetree.pattern -> Astlib.Ast_500.Parsetree.pattern

val ppat_unpack : loc:Astlib.Location.t -> string option Astlib.Location.loc -> Astlib.Ast_500.Parsetree.pattern

val ppat_exception : loc:Astlib.Location.t -> Astlib.Ast_500.Parsetree.pattern -> Astlib.Ast_500.Parsetree.pattern

val ppat_extension : loc:Astlib.Location.t -> (string Astlib.Location.loc * Astlib.Ast_500.Parsetree.payload) -> Astlib.Ast_500.Parsetree.pattern

val ppat_open : loc:Astlib.Location.t -> Astlib.Longident.t Astlib.Location.loc -> Astlib.Ast_500.Parsetree.pattern -> Astlib.Ast_500.Parsetree.pattern

val position : fname:string -> lnum:int -> bol:int -> cnum:int -> Lexing.position

val rtag : loc:Astlib.Location.t -> string Astlib.Location.loc -> bool -> Astlib.Ast_500.Parsetree.core_type list -> Astlib.Ast_500.Parsetree.row_field

val rinherit : loc:Astlib.Location.t -> Astlib.Ast_500.Parsetree.core_type -> Astlib.Ast_500.Parsetree.row_field

val psig_value : loc:Astlib.Location.t -> Astlib.Ast_500.Parsetree.value_description -> Astlib.Ast_500.Parsetree.signature_item

val psig_type : loc:Astlib.Location.t -> Astlib.Ast_500.Asttypes.rec_flag -> Astlib.Ast_500.Parsetree.type_declaration list -> Astlib.Ast_500.Parsetree.signature_item

val psig_typesubst : loc:Astlib.Location.t -> Astlib.Ast_500.Parsetree.type_declaration list -> Astlib.Ast_500.Parsetree.signature_item

val psig_typext : loc:Astlib.Location.t -> Astlib.Ast_500.Parsetree.type_extension -> Astlib.Ast_500.Parsetree.signature_item

val psig_exception : loc:Astlib.Location.t -> Astlib.Ast_500.Parsetree.type_exception -> Astlib.Ast_500.Parsetree.signature_item

val psig_module : loc:Astlib.Location.t -> Astlib.Ast_500.Parsetree.module_declaration -> Astlib.Ast_500.Parsetree.signature_item

val psig_modsubst : loc:Astlib.Location.t -> Astlib.Ast_500.Parsetree.module_substitution -> Astlib.Ast_500.Parsetree.signature_item

val psig_recmodule : loc:Astlib.Location.t -> Astlib.Ast_500.Parsetree.module_declaration list -> Astlib.Ast_500.Parsetree.signature_item

val psig_modtype : loc:Astlib.Location.t -> Astlib.Ast_500.Parsetree.module_type_declaration -> Astlib.Ast_500.Parsetree.signature_item

val psig_modtypesubst : loc:Astlib.Location.t -> Astlib.Ast_500.Parsetree.module_type_declaration -> Astlib.Ast_500.Parsetree.signature_item

val psig_open : loc:Astlib.Location.t -> Astlib.Longident.t Astlib.Location.loc Astlib.Ast_500.Parsetree.open_infos -> Astlib.Ast_500.Parsetree.signature_item

val psig_class : loc:Astlib.Location.t -> Astlib.Ast_500.Parsetree.class_type Astlib.Ast_500.Parsetree.class_infos list -> Astlib.Ast_500.Parsetree.signature_item

val psig_class_type : loc:Astlib.Location.t -> Astlib.Ast_500.Parsetree.class_type Astlib.Ast_500.Parsetree.class_infos list -> Astlib.Ast_500.Parsetree.signature_item

val psig_attribute : loc:Astlib.Location.t -> Astlib.Ast_500.Parsetree.attribute -> Astlib.Ast_500.Parsetree.signature_item

val psig_extension : loc:Astlib.Location.t -> (string Astlib.Location.loc * Astlib.Ast_500.Parsetree.payload) -> Astlib.Ast_500.Parsetree.attribute list -> Astlib.Ast_500.Parsetree.signature_item

val psig_kind_abbrev : loc:Astlib.Location.t -> string Astlib.Location.loc -> Astlib.Ast_500.Parsetree.jkind_annotation -> Astlib.Ast_500.Parsetree.signature_item

val pstr_eval : loc:Astlib.Location.t -> Astlib.Ast_500.Parsetree.expression -> Astlib.Ast_500.Parsetree.attribute list -> Astlib.Ast_500.Parsetree.structure_item

val pstr_value : loc:Astlib.Location.t -> Astlib.Ast_500.Asttypes.rec_flag -> Astlib.Ast_500.Parsetree.value_binding list -> Astlib.Ast_500.Parsetree.structure_item

val pstr_primitive : loc:Astlib.Location.t -> Astlib.Ast_500.Parsetree.value_description -> Astlib.Ast_500.Parsetree.structure_item

val pstr_type : loc:Astlib.Location.t -> Astlib.Ast_500.Asttypes.rec_flag -> Astlib.Ast_500.Parsetree.type_declaration list -> Astlib.Ast_500.Parsetree.structure_item

val pstr_typext : loc:Astlib.Location.t -> Astlib.Ast_500.Parsetree.type_extension -> Astlib.Ast_500.Parsetree.structure_item

val pstr_exception : loc:Astlib.Location.t -> Astlib.Ast_500.Parsetree.type_exception -> Astlib.Ast_500.Parsetree.structure_item

val pstr_module : loc:Astlib.Location.t -> Astlib.Ast_500.Parsetree.module_binding -> Astlib.Ast_500.Parsetree.structure_item

val pstr_recmodule : loc:Astlib.Location.t -> Astlib.Ast_500.Parsetree.module_binding list -> Astlib.Ast_500.Parsetree.structure_item

val pstr_modtype : loc:Astlib.Location.t -> Astlib.Ast_500.Parsetree.module_type_declaration -> Astlib.Ast_500.Parsetree.structure_item

val pstr_open : loc:Astlib.Location.t -> Astlib.Ast_500.Parsetree.module_expr Astlib.Ast_500.Parsetree.open_infos -> Astlib.Ast_500.Parsetree.structure_item

val pstr_class : loc:Astlib.Location.t -> Astlib.Ast_500.Parsetree.class_expr Astlib.Ast_500.Parsetree.class_infos list -> Astlib.Ast_500.Parsetree.structure_item

val pstr_class_type : loc:Astlib.Location.t -> Astlib.Ast_500.Parsetree.class_type Astlib.Ast_500.Parsetree.class_infos list -> Astlib.Ast_500.Parsetree.structure_item

val pstr_include : loc:Astlib.Location.t -> Astlib.Ast_500.Parsetree.module_expr Astlib.Ast_500.Parsetree.include_infos -> Astlib.Ast_500.Parsetree.structure_item

val pstr_attribute : loc:Astlib.Location.t -> Astlib.Ast_500.Parsetree.attribute -> Astlib.Ast_500.Parsetree.structure_item

val pstr_extension : loc:Astlib.Location.t -> (string Astlib.Location.loc * Astlib.Ast_500.Parsetree.payload) -> Astlib.Ast_500.Parsetree.attribute list -> Astlib.Ast_500.Parsetree.structure_item

val pstr_kind_abbrev : loc:Astlib.Location.t -> string Astlib.Location.loc -> Astlib.Ast_500.Parsetree.jkind_annotation -> Astlib.Ast_500.Parsetree.structure_item

val toplevel_directive : loc:Astlib.Location.t -> name:string Astlib.Location.loc -> arg:Astlib.Ast_500.Parsetree.directive_argument option -> Astlib.Ast_500.Parsetree.toplevel_directive

val type_exception : loc:Astlib.Location.t -> Astlib.Ast_500.Parsetree.extension_constructor -> Astlib.Ast_500.Parsetree.type_exception

val type_extension : loc:Astlib.Location.t -> path:Astlib.Longident.t Astlib.Location.loc -> params: (Astlib.Ast_500.Parsetree.core_type * (Astlib.Ast_500.Asttypes.variance * Astlib.Ast_500.Asttypes.injectivity)) list -> constructors:Astlib.Ast_500.Parsetree.extension_constructor list -> private_:Astlib.Ast_500.Asttypes.private_flag -> Astlib.Ast_500.Parsetree.type_extension

val pexp_apply : loc:Astlib.Location.t -> Astlib.Ast_500.Parsetree.expression -> (Astlib.Ast_500.Asttypes.arg_label * Astlib.Ast_500.Parsetree.expression) list -> Astlib.Ast_500.Parsetree.expression

val ppat_construct : loc:Astlib.Location.t -> Astlib.Longident.t Astlib.Location.loc -> Astlib.Ast_500.Parsetree.pattern option -> Astlib.Ast_500.Parsetree.pattern

val constructor_declaration : loc:Astlib.Location.t -> name:string Astlib.Location.loc -> args:Astlib.Ast_500.Parsetree.constructor_arguments -> res:Astlib.Ast_500.Parsetree.core_type option -> Astlib.Ast_500.Parsetree.constructor_declaration

val eint : loc:Ppxlib.Location.t -> int -> Astlib.Ast_500.Parsetree.expression

val echar : loc:Ppxlib.Location.t -> char -> Astlib.Ast_500.Parsetree.expression

val estring : loc:Ppxlib.Location.t -> string -> Astlib.Ast_500.Parsetree.expression

val efloat : loc:Ppxlib.Location.t -> string -> Astlib.Ast_500.Parsetree.expression

val eint32 : loc:Ppxlib.Location.t -> int32 -> Astlib.Ast_500.Parsetree.expression

val eint64 : loc:Ppxlib.Location.t -> int64 -> Astlib.Ast_500.Parsetree.expression

val enativeint : loc:Ppxlib.Location.t -> nativeint -> Astlib.Ast_500.Parsetree.expression

val ebool : loc:Ppxlib.Location.t -> bool -> Astlib.Ast_500.Parsetree.expression

val pint : loc:Ppxlib.Location.t -> int -> Astlib.Ast_500.Parsetree.pattern

val pchar : loc:Ppxlib.Location.t -> char -> Astlib.Ast_500.Parsetree.pattern

val pstring : loc:Ppxlib.Location.t -> string -> Astlib.Ast_500.Parsetree.pattern

val pfloat : loc:Ppxlib.Location.t -> string -> Astlib.Ast_500.Parsetree.pattern

val pint32 : loc:Ppxlib.Location.t -> int32 -> Astlib.Ast_500.Parsetree.pattern

val pint64 : loc:Ppxlib.Location.t -> int64 -> Astlib.Ast_500.Parsetree.pattern

val pnativeint : loc:Ppxlib.Location.t -> nativeint -> Astlib.Ast_500.Parsetree.pattern

val pbool : loc:Ppxlib.Location.t -> bool -> Astlib.Ast_500.Parsetree.pattern

val eunit : loc:Ppxlib.Location.t -> Astlib.Ast_500.Parsetree.expression

val punit : loc:Ppxlib.Location.t -> Astlib.Ast_500.Parsetree.pattern

val evar : loc:Ppxlib.Location.t -> string -> Astlib.Ast_500.Parsetree.expression

evar id produces a Pexp_ident _ expression, it parses its input so you can pass any dot-separated identifier, for instance: evar ~loc "Foo.bar".

val pvar : loc:Ppxlib.Location.t -> string -> Astlib.Ast_500.Parsetree.pattern

val eapply : loc:Ppxlib.Location.t -> Astlib.Ast_500.Parsetree.expression -> Astlib.Ast_500.Parsetree.expression list -> Astlib.Ast_500.Parsetree.expression

Same as pexp_apply but without labels

val esequence : loc:Ppxlib.Location.t -> Astlib.Ast_500.Parsetree.expression list -> Astlib.Ast_500.Parsetree.expression

val ppat_tuple_opt : loc:Ppxlib.Location.t -> Astlib.Ast_500.Parsetree.pattern list -> Astlib.Ast_500.Parsetree.pattern option

val pexp_tuple_opt : loc:Ppxlib.Location.t -> Astlib.Ast_500.Parsetree.expression list -> Astlib.Ast_500.Parsetree.expression option

val pconstruct : Astlib.Ast_500.Parsetree.constructor_declaration -> Astlib.Ast_500.Parsetree.pattern option -> Astlib.Ast_500.Parsetree.pattern

val econstruct : Astlib.Ast_500.Parsetree.constructor_declaration -> Astlib.Ast_500.Parsetree.expression option -> Astlib.Ast_500.Parsetree.expression

val elist : loc:Ppxlib.Location.t -> Astlib.Ast_500.Parsetree.expression list -> Astlib.Ast_500.Parsetree.expression

val plist : loc:Ppxlib.Location.t -> Astlib.Ast_500.Parsetree.pattern list -> Astlib.Ast_500.Parsetree.pattern

val pexp_function : loc:Ppxlib.Location.t -> Astlib.Ast_500.Parsetree.case list -> Astlib.Ast_500.Parsetree.expression

val pexp_fun : loc:Ppxlib.Location.t -> Astlib.Ast_500.Asttypes.arg_label -> Astlib.Ast_500.Parsetree.expression option -> Astlib.Ast_500.Parsetree.pattern -> Astlib.Ast_500.Parsetree.expression -> Astlib.Ast_500.Parsetree.expression

val pstr_value_list : loc:Ppxlib.Location.t -> Ppxlib_ast.Asttypes.rec_flag -> Astlib.Ast_500.Parsetree.value_binding list -> Astlib.Ast_500.Parsetree.structure_item list

pstr_value_list ~loc rf vbs = pstr_value ~loc rf vbs if vbs <> [], [] otherwise.

val nonrec_type_declaration : loc:Ppxlib.Location.t -> name:string Ppxlib.Loc.t -> params: (Astlib.Ast_500.Parsetree.core_type * Ppxlib_ast.Asttypes.variance) list -> cstrs: (Astlib.Ast_500.Parsetree.core_type * Astlib.Ast_500.Parsetree.core_type * Ppxlib.Location.t) list -> kind:Astlib.Ast_500.Parsetree.type_kind -> private_:Ppxlib_ast.Asttypes.private_flag -> manifest:Astlib.Ast_500.Parsetree.core_type option -> Astlib.Ast_500.Parsetree.type_declaration

deprecated [since 2016-10] use Nonrecursive on the P(str|sig)_type instead

val unapplied_type_constr_conv : loc:Ppxlib.Location.t -> Ppxlib.Longident.t Ppxlib.Loc.t -> f:(string -> string) -> Astlib.Ast_500.Parsetree.expression

unapplied_type_constr_conv is the standard way to map identifiers to conversion fonctions, for preprocessor that creates values that follow the structure of types. More precisely, path_conv path (sprintf "sexp_of_%s") is:

sexp_of_t if path is "t"

A.B.sexp_of_foo if path is "A.B.foo"

A.B.sexp_of_f__foo (module A1) (module A2) if path is "A.B.F(A1)(A2).foo" type_constr_conv also applies it to a list of expression, which both prevents the compiler from allocating useless closures, and almost always what is needed, since type constructors are always applied.

val type_constr_conv : loc:Ppxlib.Location.t -> Ppxlib.Longident.t Ppxlib.Loc.t -> f:(string -> string) -> Astlib.Ast_500.Parsetree.expression list -> Astlib.Ast_500.Parsetree.expression

val eta_reduce : Astlib.Ast_500.Parsetree.expression -> Astlib.Ast_500.Parsetree.expression option

Tries to simplify fun v1 v2 .. -> f v1 v2 .. into f. Only works when f is a path, not an arbitrary expression as that would change the meaning of the code. This can be used either for cleaning up the generated code, or to reduce allocation if f is a local variable (the compiler won't optimize the allocation of the closure).

Eta-reduction can change the types/behavior in some corner cases that are unlikely to show up in generated code:

if f has optional arguments, eta-expanding f can drop them

because labels commute, it can change the type of an expression: $ let f ~x y = x + y let f2 = fun x -> add x;; val f : x:int -> int -> int = <fun> val f2 : int -> x:int -> int = <fun> In fact, if f does side effects before receiving all its arguments, and if the eta-expansion is partially applied, eta-reducing could change behavior.

eta_reduce_if_possible_and_nonrec is meant for the case where the resulting expression is going to be bound in a potentially recursive let-binding, where we have to keep the eta-expansion when rec_flag is Recursive to avoid a compile error.

val eta_reduce_if_possible : Astlib.Ast_500.Parsetree.expression -> Astlib.Ast_500.Parsetree.expression

val eta_reduce_if_possible_and_nonrec : Astlib.Ast_500.Parsetree.expression -> rec_flag:Astlib.Ast_500.Asttypes.rec_flag -> Astlib.Ast_500.Parsetree.expression

val ptyp_any : loc:Ppxlib.Location.t -> Astlib.Ast_500.Parsetree.core_type

val ptyp_var : loc:Ppxlib.Location.t -> string -> Astlib.Ast_500.Parsetree.core_type

include module type of struct include Ppxlib_jane.Ast_builder.Default end

type index_kind = Ppxlib_jane.Shim.index_kind

Modes

val ptyp_arrow : loc:Astlib.Location.t -> Ppxlib_jane.Shim.arrow_argument -> Ppxlib_jane.Shim.arrow_result -> Ppxlib_ast.Parsetree.core_type

Construct a Ptyp_arrow

val tarrow : loc:Astlib.Location.t -> Ppxlib_jane.Shim.arrow_argument list -> Ppxlib_jane.Shim.arrow_result -> Ppxlib_ast.Parsetree.core_type

Construct a multi-argument arrow type with the provided arguments and result.

raises [Invalid_argument]
if the input list is empty.

val tarrow_maybe : loc:Astlib.Location.t -> Ppxlib_jane.Shim.arrow_argument list -> Ppxlib_ast.Parsetree.core_type -> Ppxlib_ast.Parsetree.core_type

As tarrow, but will return the result if the input list is empty rather than erroring; this means the result type cannot have a mode annotation.

val pexp_let : loc:Astlib.Location.t -> Ppxlib_ast.Asttypes.mutable_flag -> Ppxlib_ast.Asttypes.rec_flag -> Ppxlib_ast.Parsetree.value_binding list -> Ppxlib_ast.Parsetree.expression -> Ppxlib_ast.Parsetree.expression

Pexp_let with let mutable support

val pexp_constraint : loc:Astlib.Location.t -> Ppxlib_ast.Parsetree.expression -> Ppxlib_ast.Parsetree.core_type option -> Ppxlib_jane.Shim.Modes.t -> Ppxlib_ast.Parsetree.expression

Construct a Pexp_constraint with modes

val ppat_constraint : loc:Astlib.Location.t -> Ppxlib_ast.Parsetree.pattern -> Ppxlib_ast.Parsetree.core_type option -> Ppxlib_jane.Shim.Modes.t -> Ppxlib_ast.Parsetree.pattern

Construct a Ppat_constraint with modes

val value_binding : loc:Astlib.Location.t -> pat:Ppxlib_ast.Parsetree.pattern -> expr:Ppxlib_ast.Parsetree.expression -> modes:Ppxlib_jane.Shim.Modes.t -> Ppxlib_ast.Parsetree.value_binding

Contruct a value_binding with modes

val pcstr_tuple : loc:Astlib.Location.t -> (Ppxlib_jane.Shim.Modalities.t * Ppxlib_ast.Parsetree.core_type) list -> Ppxlib_ast.Parsetree.constructor_arguments

Construct a Pcstr_tuple, a representation for the contents of a tupled variant constructor, that attaches the provided modalities to each field.

val psig_include : loc:Astlib.Location.t -> modalities:Ppxlib_jane.Shim.Modalities.t -> Ppxlib_ast.Parsetree.include_description -> Ppxlib_ast.Parsetree.signature_item

Construct a Psig_include with modalities

val signature : loc:Astlib.Location.t -> ?modalities:Ppxlib_jane.Shim.Modalities.t -> Ppxlib_ast.Parsetree.signature_item list -> Ppxlib_ast.Parsetree.signature

Construct a signature

val pmty_signature : loc:Astlib.Location.t -> Ppxlib_ast.Parsetree.signature -> Ppxlib_ast.Parsetree.module_type

val get_tuple_field_modalities : Ppxlib_jane.Shim.Pcstr_tuple_arg.t -> Ppxlib_jane.Shim.Modalities.t * Ppxlib_ast.Parsetree.core_type

Splits a possibly-modality-annotated field of a tupled variant constructor into a pair of its modality and the unannotated field. If the resulting mode is None, then the field is returned unchanged.

val get_label_declaration_modalities : Ppxlib_ast.Parsetree.label_declaration -> Ppxlib_jane.Shim.Modalities.t * Ppxlib_ast.Parsetree.label_declaration

Splits a possibly-modality-annotated label declaration into a pair of its modality and the unannotated label declaration. If the resulting modality is None, then the label declaration is returned unchanged.

val label_declaration : loc:Astlib.Location.t -> name:string Astlib.Location.loc -> mutable_:Ppxlib_ast.Asttypes.mutable_flag -> modalities:Ppxlib_jane.Shim.Modalities.t -> type_:Ppxlib_ast.Parsetree.core_type -> Ppxlib_ast.Parsetree.label_declaration

val get_value_description_modalities : Ppxlib_ast.Parsetree.value_description -> Ppxlib_jane.Shim.Modalities.t * Ppxlib_ast.Parsetree.value_description

val value_description : loc:Astlib.Location.t -> name:string Astlib.Location.loc -> type_:Ppxlib_ast.Parsetree.core_type -> modalities:Ppxlib_jane.Shim.Modalities.t -> prim:string list -> Ppxlib_ast.Parsetree.value_description

val pcstr_tuple_arg : loc:Astlib.Location.t -> modalities:Ppxlib_jane.Shim.Modalities.t -> type_:Ppxlib_ast.Parsetree.core_type -> Ppxlib_jane.Shim.Pcstr_tuple_arg.t

val include_infos : loc:Astlib.Location.t -> ?attrs:Ppxlib_ast.Parsetree.attributes -> kind:Ppxlib_jane.Shim.Include_kind.t -> 'a -> 'a Ppxlib_ast.Parsetree.include_infos

val module_declaration : loc:Astlib.Location.t -> ?attrs:Ppxlib_ast.Parsetree.attributes -> ?modalities:Ppxlib_jane.Shim.Modalities.t -> string option Astlib.Location.loc -> Ppxlib_ast.Parsetree.module_type -> Ppxlib_ast.Parsetree.module_declaration

val pmty_functor : loc:Astlib.Location.t -> ?attrs:Ppxlib_ast.Parsetree.attributes -> ?modes:Ppxlib_jane.Shim.Modes.t -> Ppxlib_ast.Parsetree.functor_parameter -> Ppxlib_ast.Parsetree.module_type -> Ppxlib_ast.Parsetree.module_type

val pmod_constraint : loc:Astlib.Location.t -> ?attrs:Ppxlib_ast.Parsetree.attributes -> Ppxlib_ast.Parsetree.module_expr -> Ppxlib_ast.Parsetree.module_type option -> Ppxlib_jane.Shim.Modes.t -> Ppxlib_ast.Parsetree.module_expr

N-ary functions

Many comments below make reference to the Jane Street compiler's treatment of function arity. These comments refer to a parsetree change made to upstream OCaml in https://github.com/ocaml/ocaml/pull/12236, but that Jane Street has mirrored internally already.

The treatment of arity can be summarized as follows:

In a previous version of OCaml, a function's runtime arity was inferred at a late stage of the compiler, after typechecking, where it fuses together nested lambdas.

In the new version of OCaml (both upstream OCaml after #12236 and the internal Jane Street compiler), a function's runtime arity is purely a syntactic notion: it's the number of parameters in a fun x1 ... xn -> body construct, with some special allowances for function cases.

Why is arity important? In native code, application sites of a function to n syntactic arguments will trigger a fast path (where arguments are passed in registers) only if the function's runtime arity is n.

As a result, ppxes must take more care than before to generate functions of the correct arity. Now, a nested function like fun x -> fun y -> e has arity 1 (returning still another function of arity 1) instead of arity 2. All bindings below that construct functions are documented as to the arity of the returned function.

Some examples of arity:

2-ary function: fun x y -> e

1-ary function returning 1-ary function: fun x -> fun y -> e

3-ary function: fun x y -> function P1 -> e1 | P2 -> e2

2-ary function returning 1-ary function: fun x y -> (function P1 -> e1 | P2 -> e2)

2-ary function returning 1-ary function: fun x -> function P1 -> function P2 -> e

Notably, unparenthesized function has a special meaning when used as a direct body of fun: the function becomes part of the arity of the outer fun. The same does not apply for multiple nested functions, even if they each have a single case; the nested functions are treated as unary. (See the last example.)

type function_param = Ppxlib_jane.Shim.Pexp_function.function_param

type function_constraint = Ppxlib_jane.Shim.Pexp_function.Function_constraint.t

type function_body = Ppxlib_jane.Shim.Pexp_function.function_body

module Latest = Ppxlib_jane.Ast_builder.Default.Latest

val eabstract : loc:Astlib.Location.t -> ?coalesce_fun_arity:bool -> ?return_constraint:Ppxlib_ast.Parsetree.core_type -> Ppxlib_ast.Parsetree.pattern list -> Ppxlib_ast.Parsetree.expression -> Ppxlib_ast.Parsetree.expression

Create a function with unlabeled parameters and an expression body. Like Ppxlib.Ast_builder.eapply, but for constructing functions.

coalesce_fun_arity is relevant for the Jane Street compiler. By default, coalesce_fun_arity is true.

Suppose there is a call eabstract pats body ~coalesce_fun_arity

If colaesce_fun_arity is true, the arity of the returned function is the same as the arity of: add_fun_params (List.map params ~f:(Fun.param Nolabel)) body

If coalesce_fun_arity is false, then the arity of the returned function is the length of pats.

In other words, coalesce_fun_arity = true allows you to build up the arity of an already-constructed function rather than necessarily creating a new function.

val unary_function : loc:Astlib.Location.t -> ?attrs:Ppxlib_ast.Parsetree.attributes -> Ppxlib_ast.Parsetree.case list -> Ppxlib_ast.Parsetree.expression

unary_function cases is function <cases>. When used with the Jane Street compiler, the function's runtime arity is 1, so the fast path for function application happens only when application sites of the resulting function receive 1 argument. To create a function with multiple argument that pattern-matches on the last one, use add_param or add_params to add more parameters. Alternatively, use pexp_function to provide all parameters at once.

The attributes of the resulting expression will be the attrs argument together with any attributes added by the Jane Street compiler.

val fun_param : loc:Astlib.Location.t -> Ppxlib_ast.Asttypes.arg_label -> Ppxlib_ast.Parsetree.pattern -> function_param

fun_param lbl pat is Pparam_val (lbl, None, pat). This gives a more self-documenting way of constructing the usual case: value parameters without optional argument defaults.

val add_fun_param : loc:Astlib.Location.t -> ?attrs:Ppxlib_ast.Parsetree.attributes -> ?return_constraint:Ppxlib_ast.Parsetree.core_type -> Ppxlib_ast.Asttypes.arg_label -> Ppxlib_ast.Parsetree.expression option -> Ppxlib_ast.Parsetree.pattern -> Ppxlib_ast.Parsetree.expression -> Ppxlib_ast.Parsetree.expression

Say an expression is a "function" if it is a Pexp_fun or a Pexp_function. All functions have parameters and arity.

Suppose add_param lbl def pat e ==> e'. Then, letting param = Pparam_val (lbl, def, pat),

If e is a function with arity n, then e' is a function with arity n+1. param is added at the outermost layer. For example, if e = fun <params> -> <body>, then e' = fun <param :: params> -> body. The attributes on the resulting expression will be the attrs argument together with any attributes already present on e.

If e is not a function, then e' is a function with arity 1, namely: fun <param> -> <e>. The attributes of the resulting expression will be the attrs argument together with any attributes added by the Jane Street compiler.

val add_fun_params : loc:Astlib.Location.t -> ?attrs:Ppxlib_ast.Parsetree.attributes -> ?return_constraint:Ppxlib_ast.Parsetree.core_type -> function_param list -> Ppxlib_ast.Parsetree.expression -> Ppxlib_ast.Parsetree.expression

add_params params e is List.fold_right params ~init:e ~f:add_param. Note the fold_right: if e is fun <params'> -> <body>, then add_params params e is fun <params @ params'> -> <body>.

val coalesce_fun_arity : Ppxlib_ast.Parsetree.expression -> Ppxlib_ast.Parsetree.expression

This operation is a no-op, except as interpreted by the Jane Street compiler. If e is a function with arity n with an expression body that itself is a function with arity m, then coalesce_fun_arity e is a function of arity n + m.

You should usually call coalesce_fun_arity on metaquot fun expressions whose body may be a function, e.g.:

coalesce_fun_arity [%expr fun x y -> [%e possibly_function]]

Unboxed types

val ptyp_unboxed_tuple : loc:Astlib.Location.t -> (string option * Ppxlib_ast.Parsetree.core_type) list -> Ppxlib_ast.Parsetree.core_type

val pexp_unboxed_tuple : loc:Astlib.Location.t -> ?attrs:Ppxlib_ast.Parsetree.attributes -> (string option * Ppxlib_ast.Parsetree.expression) list -> Ppxlib_ast.Parsetree.expression

val pexp_record_unboxed_product : loc:Astlib.Location.t -> ?attrs:Ppxlib_ast.Parsetree.attributes -> (Astlib.Longident.t Ppxlib_ast.Asttypes.loc * Ppxlib_ast.Parsetree.expression) list -> Ppxlib_ast.Parsetree.expression option -> Ppxlib_ast.Parsetree.expression

val ppat_record_unboxed_product : loc:Astlib.Location.t -> ?attrs:Ppxlib_ast.Parsetree.attributes -> (Astlib.Longident.t Ppxlib_ast.Asttypes.loc * Ppxlib_ast.Parsetree.pattern) list -> Ppxlib_ast.Asttypes.closed_flag -> Ppxlib_ast.Parsetree.pattern

val pexp_unboxed_field : loc:Astlib.Location.t -> ?attrs:Ppxlib_ast.Parsetree.attributes -> Ppxlib_ast.Parsetree.expression -> Astlib.Longident.t Ppxlib_ast.Asttypes.loc -> Ppxlib_ast.Parsetree.expression

val ppat_unboxed_tuple : loc:Astlib.Location.t -> ?attrs:Ppxlib_ast.Parsetree.attributes -> (string option * Ppxlib_ast.Parsetree.pattern) list -> Ppxlib_ast.Asttypes.closed_flag -> Ppxlib_ast.Parsetree.pattern

val pexp_idx : loc:Astlib.Location.t -> ?attrs:Ppxlib_ast.Parsetree.attributes -> Ppxlib_jane.Shim.block_access -> Ppxlib_jane.Shim.unboxed_access list -> Ppxlib_ast.Parsetree.expression

Expression literals

val eint64_u : loc:Astlib.Location.t -> int64 -> Ppxlib_ast.Parsetree.expression

e.g. #42L

val eint32_u : loc:Astlib.Location.t -> int32 -> Ppxlib_ast.Parsetree.expression

e.g. #42l

val enativeint_u : loc:Astlib.Location.t -> nativeint -> Ppxlib_ast.Parsetree.expression

e.g. #42n

val efloat_u : loc:Astlib.Location.t -> string -> Ppxlib_ast.Parsetree.expression

e.g. #42.

Pattern literals

val pint64_u : loc:Astlib.Location.t -> int64 -> Ppxlib_ast.Parsetree.pattern

e.g. #42L

val pint32_u : loc:Astlib.Location.t -> int32 -> Ppxlib_ast.Parsetree.pattern

e.g. #42l

val pnativeint_u : loc:Astlib.Location.t -> nativeint -> Ppxlib_ast.Parsetree.pattern

e.g. #42n

val pfloat_u : loc:Astlib.Location.t -> string -> Ppxlib_ast.Parsetree.pattern

e.g. #42.

Layouts

val ptyp_alias : loc:Astlib.Location.t -> ?attrs:Ppxlib_ast.Parsetree.attributes -> Ppxlib_ast.Parsetree.core_type -> Ppxlib_ast.Asttypes.label Ppxlib_ast.Asttypes.loc option -> Ppxlib_jane.Shim.jkind_annotation option -> Ppxlib_ast.Parsetree.core_type

val ptyp_poly : loc:Astlib.Location.t -> ?attrs:Ppxlib_ast.Parsetree.attributes -> (string Ppxlib_ast.Asttypes.loc * Ppxlib_jane.Shim.jkind_annotation option) list -> Ppxlib_ast.Parsetree.core_type -> Ppxlib_ast.Parsetree.core_type

val ptyp_quote : loc:Astlib.Location.t -> ?attrs:Ppxlib_ast.Parsetree.attributes -> Ppxlib_ast.Parsetree.core_type -> Ppxlib_ast.Parsetree.core_type

val ptyp_splice : loc:Astlib.Location.t -> ?attrs:Ppxlib_ast.Parsetree.attributes -> Ppxlib_ast.Parsetree.core_type -> Ppxlib_ast.Parsetree.core_type

val pexp_newtype : loc:Astlib.Location.t -> ?attrs:Ppxlib_ast.Parsetree.attributes -> string Ppxlib_ast.Asttypes.loc -> Ppxlib_jane.Shim.jkind_annotation option -> Ppxlib_ast.Parsetree.expression -> Ppxlib_ast.Parsetree.expression

val type_declaration : loc:Astlib.Location.t -> ?attrs:Ppxlib_ast.Parsetree.attributes -> ?jkind_annotation:Ppxlib_jane.Shim.jkind_annotation -> name:string Astlib.Location.loc -> params: (Ppxlib_ast.Parsetree.core_type * (Ppxlib_ast.Asttypes.variance * Ppxlib_ast.Asttypes.injectivity)) list -> cstrs: (Ppxlib_ast.Parsetree.core_type * Ppxlib_ast.Parsetree.core_type * Astlib.Location.t) list -> kind:Ppxlib_ast.Parsetree.type_kind -> private_:Ppxlib_ast.Asttypes.private_flag -> manifest:Ppxlib_ast.Parsetree.core_type option -> unit -> Ppxlib_ast.Parsetree.type_declaration

Labeled tuples

val ptyp_tuple : loc:Astlib.Location.t -> ?attrs:Ppxlib_ast.Parsetree.attributes -> (string option * Ppxlib_ast.Parsetree.core_type) list -> Ppxlib_ast.Parsetree.core_type

val pexp_tuple : loc:Astlib.Location.t -> ?attrs:Ppxlib_ast.Parsetree.attributes -> (string option * Ppxlib_ast.Parsetree.expression) list -> Ppxlib_ast.Parsetree.expression

val ppat_tuple : loc:Astlib.Location.t -> ?attrs:Ppxlib_ast.Parsetree.attributes -> (string option * Ppxlib_ast.Parsetree.pattern) list -> Ppxlib_ast.Asttypes.closed_flag -> Ppxlib_ast.Parsetree.pattern

Immutable arrays

val ppat_array : loc:Astlib.Location.t -> ?attrs:Ppxlib_ast.Parsetree.attributes -> Ppxlib_ast.Asttypes.mutable_flag -> Ppxlib_ast.Parsetree.pattern list -> Ppxlib_ast.Parsetree.pattern

val pexp_array : loc:Astlib.Location.t -> ?attrs:Ppxlib_ast.Parsetree.attributes -> Ppxlib_ast.Asttypes.mutable_flag -> Ppxlib_ast.Parsetree.expression list -> Ppxlib_ast.Parsetree.expression